Don't use latest=>true for commits and tags
The Puppet git_resource provider implementation makes "latest" a
no-op when using a commit or tag reference. This may cause it to
fail to update an existing worktree to a newer ref because Puppet
will check the "latest" method and assume nothing needs to be done.
Drop the latest=>true options from those conditional cases,
retaining it only for branches/heads.

Change-Id: I8986c32e6cbf2eca0e0624b014f0fca08b04f145