Don't require virtualenv package
The puppet-python does a check for the virtualenv package, and thus
declares it's own virtualenv Package ... we do the "if ! defined" but
with ordering changes this can conflict up the upstream module.

Also, puppet-python by default ensure the -dev package is absent.
Ensure this doesn't happen as we need to build some pip installed
packages.

Depends-On: https://review.openstack.org/559871
Change-Id: I3435a2527f4712d068c5a3c43728bed6011e1ebd