Fix template variable
I forgot that on puppet 3, scope.lookupvar returns :undef instead of nil
when it looks up a variable set to 'undef'. Since we're not setting the
chain file in our beaker tests, this causes an invalid apache config to
be rendered. This patch ensures the file explicitly checks for both nil
and :undef. On puppet 4, checking for just nil will be sufficient.

Change-Id: I7fd806c65f1aef95618a4549b24c6fa21aa3fe29