Fix apache compatibility for Xenial
Without this patch, the askbot vhost template compares the
lsbdistcodename fact with the string 'trusty' to decide whether to use
the Apache 2.4 syntax, and since 'xenial' does not match 'trusty' it was
falling back to the 2.2 syntax, causing invalid apache configuration and
resulting in 403 errors for the site. Instead of keeping this logic in
puppet/ERB, this patch updates the template to use the apache
mod_version module to decide what syntax to use.

With this working, we can add curl test to ensure the site functions
properly.

Depends-On: https://review.openstack.org/585195

Change-Id: I3bc4a2d96ea6483119a504a1d50bf92357de254d